Split step() into two other methods + refactor
The MailGenerator.step() method was doing too many things and the name
wasn't capable of properly describing it. This is usually a big sign
that a method has too many responsibilities and should be split up, in
order to increase cohesion. Additionally, mailpool is now passed in as
arguments, rather than MailGenerator having it as a field. This is
because it was one method anyway and passing it in as an argument is
more modular. This means that MailPool has been removed from
MailGenerator's constructor.